- Added a psprite flag for the sprites section in a wl6map.  This changes the offsets ECWolf uses during the conversion. - Fixed: Missing column in wolf shapes. - Player sprite drawing is done extremely similarly to ZDoom.  This means high resolution weapons are now supported and in fact you could take a sprite from Doom and use it as-is. - Did some work on supporting decorate weapons.

+++ src/m_swap.h
@@ -51,7 +51,7 @@ static inline const WORD SwapShort(WORD x)
 	return ((x&0xFF)<<8) | ((x>>8)&0xFF);
 }
 
-static inline const WORD SwapLong(DWORD x)
+static inline const DWORD SwapLong(DWORD x)
 {
 	return ((x&0xFF)<<24) |
 		(((x>>8)&0xFF)<<16) |
